+++ crypto/coconut/concurrency/coconutworker/coconut_worker.go
@@ -21,79 +21,80 @@ package coconutworker
 import (
 	"sync"
 
+	"0xacab.org/jstuczyn/CoconutGo/crypto/coconut/concurrency/jobpacket"
 	"0xacab.org/jstuczyn/CoconutGo/crypto/coconut/scheme"
 	"0xacab.org/jstuczyn/CoconutGo/crypto/elgamal"
 	Curve "github.com/jstuczyn/amcl/version3/go/amcl/BLS381"
 )
 
-// Worker allows writing coconut actions to a shared job queue,
+// CoconutWorker allows writing coconut actions to a shared job queue,
 // so that they could be run concurrently.
-type Worker struct {
-	jobQueue  chan<- interface{}
+type CoconutWorker struct {
+	jobQueue  chan<- *jobpacket.JobPacket
 	muxParams *MuxParams
 }
 
 // SignWrapper wraps the provided arguments with pre-generated params.
-func (cw *Worker) SignWrapper(sk *coconut.SecretKey, pubM []*Curve.BIG) (*coconut.Signature, error) {
+func (cw *CoconutWorker) SignWrapper(sk *coconut.SecretKey, pubM []*Curve.BIG) (*coconut.Signature, error) {
 	return cw.Sign(cw.muxParams, sk, pubM)
 }
 
 // BlindSignWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) BlindSignWrapper(sk *coconut.SecretKey, blindSignMats *coconut.BlindSignMats, egPub *elgamal.PublicKey, pubM []*Curve.BIG) (*coconut.BlindedSignature, error) {
+func (cw *CoconutWorker) BlindSignWrapper(sk *coconut.SecretKey, blindSignMats *coconut.BlindSignMats, egPub *elgamal.PublicKey, pubM []*Curve.BIG) (*coconut.BlindedSignature, error) {
 	return cw.BlindSign(cw.muxParams, sk, blindSignMats, egPub, pubM)
 }
 
 // VerifyWrapper wraps the provided arguments with pre-generated params.
-func (cw *Worker) VerifyWrapper(vk *coconut.VerificationKey, pubM []*Curve.BIG, sig *coconut.Signature) bool {
+func (cw *CoconutWorker) VerifyWrapper(vk *coconut.VerificationKey, pubM []*Curve.BIG, sig *coconut.Signature) bool {
 	return cw.Verify(cw.muxParams, vk, pubM, sig)
 }
 
 // BlindVerifyWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) BlindVerifyWrapper(vk *coconut.VerificationKey, sig *coconut.Signature, blindShowMats *coconut.BlindShowMats, pubM []*Curve.BIG) bool {
+func (cw *CoconutWorker) BlindVerifyWrapper(vk *coconut.VerificationKey, sig *coconut.Signature, blindShowMats *coconut.BlindShowMats, pubM []*Curve.BIG) bool {
 	return cw.BlindVerify(cw.muxParams, vk, sig, blindShowMats, pubM)
 }
 
 // AggregateVerificationKeysWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) AggregateVerificationKeysWrapper(vks []*coconut.VerificationKey, pp *coconut.PolynomialPoints) *coconut.VerificationKey {
+func (cw *CoconutWorker) AggregateVerificationKeysWrapper(vks []*coconut.VerificationKey, pp *coconut.PolynomialPoints) *coconut.VerificationKey {
 	return cw.AggregateVerificationKeys(cw.muxParams, vks, pp)
 }
 
 // AggregateSignaturesWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) AggregateSignaturesWrapper(sigs []*coconut.Signature, pp *coconut.PolynomialPoints) *coconut.Signature {
+func (cw *CoconutWorker) AggregateSignaturesWrapper(sigs []*coconut.Signature, pp *coconut.PolynomialPoints) *coconut.Signature {
 	return cw.AggregateSignatures(cw.muxParams, sigs, pp)
 }
 
 // UnblindWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) UnblindWrapper(blindedSignature *coconut.BlindedSignature, egPub *elgamal.PrivateKey) *coconut.Signature {
+func (cw *CoconutWorker) UnblindWrapper(blindedSignature *coconut.BlindedSignature, egPub *elgamal.PrivateKey) *coconut.Signature {
 	return cw.Unblind(cw.muxParams, blindedSignature, egPub)
 }
 
 // RandomizeWrapper wraps the provided arguments with pre-generated params.
-func (cw *Worker) RandomizeWrapper(sig *coconut.Signature) *coconut.Signature {
+func (cw *CoconutWorker) RandomizeWrapper(sig *coconut.Signature) *coconut.Signature {
 	return cw.Randomize(cw.muxParams, sig)
 }
 
 // PrepareBlindSignWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) PrepareBlindSignWrapper(egPub *elgamal.PublicKey, pubM []*Curve.BIG, privM []*Curve.BIG) (*coconut.BlindSignMats, error) {
+func (cw *CoconutWorker) PrepareBlindSignWrapper(egPub *elgamal.PublicKey, pubM []*Curve.BIG, privM []*Curve.BIG) (*coconut.BlindSignMats, error) {
 	return cw.PrepareBlindSign(cw.muxParams, egPub, pubM, privM)
 }
 
 // ShowBlindSignatureWrapper wraps the provided arguments with pre-generated params.
 // nolint: lll
-func (cw *Worker) ShowBlindSignatureWrapper(vk *coconut.VerificationKey, sig *coconut.Signature, privM []*Curve.BIG) (*coconut.BlindShowMats, error) {
+func (cw *CoconutWorker) ShowBlindSignatureWrapper(vk *coconut.VerificationKey, sig *coconut.Signature, privM []*Curve.BIG) (*coconut.BlindShowMats, error) {
 	return cw.ShowBlindSignature(cw.muxParams, vk, sig, privM)
 }
 
 // New creates new instance of a coconutWorker.
-func New(jobQueue chan<- interface{}, params *coconut.Params) *Worker {
+func New(jobQueue chan<- *jobpacket.JobPacket, params *coconut.Params) *CoconutWorker {
 	muxParams := &MuxParams{params, sync.Mutex{}}
-	cw := &Worker{
+	cw := &CoconutWorker{
 		jobQueue:  jobQueue,
 		muxParams: muxParams,
 	}