+++ docs/changelog.md
@@ -1,5 +1,25 @@
 # CoconutGo Changelog
 
+## v0.1.0
+
+* Created template to generate infinite channel behaviour for any type
+
+## v0.0.4
+
+* Reimplementation of JobQueue with different queue implementation to introduce thread safety
+
+## v0.0.3
+
+* Refactored server/comm/utils/utils.go
+* Introduced ServerMetadata struct used in ServerRequests/Responses + associated changes
+* Renamed crypto/coconut/concurrency/coconutworker/coconut_worker.go Worker to CoconutWorker + associated changes
+* Renamed client/cryptoworker/cryptoworker.go Worker to CryptoWorker + associated changes
+* Refactored /home/jedrzej/go/src/0xacab.org/jstuczyn/CoconutGo/server/cryptoworker/cryptoworker.go + associated changes
+
+## v0.0.2
+
+* Fixes jstuczyn/CoconutGo#4
+
 ## v0.0.1 - Initial Release
 
 * Coconut Signature Scheme